The evaluation function regards doubled, backward and isolated pawns

+++ include/game.hpp
@@ -21,9 +21,9 @@ public:
     std::vector<unsigned short> get_all_moves(std::string& game_result) const;
     std::vector<unsigned short> get_all_moves(const Board& board, std::string& game_result) const;
     std::string send_best_move();
-    int eval(const Board& board) const;
-    std::pair<int, unsigned short> alpha_beta(const Board& init, const int depth, int alpha,
-                                              int beta) const;
+    float eval(const Board& board) const;
+    std::pair<float, unsigned short> alpha_beta(const Board& init, const int depth, float alpha,
+                                                float beta) const;
     bool get_move(const std::string& move_str);
     void add_piece(const std::string& piece, const Color color);
     void guess_castlings();
